extract generic CANopen code
